Refactored HUD files into separate directories. 37 / 66 tests work.
