+++ src/hud/core/hudCommon.js
