+++ src/hud/overlays/cinematicHud.js
@@ -1,4 +1,4 @@
-import { createPanel, mountPanel, destroyPanel, ensureHudRoot, createButton } from "./hudCommon.js";
+import { createPanel, mountPanel, destroyPanel, ensureHudRoot, createButton } from "../core/hudCommon.js";
 
 export function createCinematicHUD({ getWorld } = {}) {
   const panel = createPanel({ id: "hud-cinematic", title: "Explosion Cinematic" });
